feat: erweitere die E-Mail-Verifizierung mit Code-Validierung und Cookie-Setzung
- Füge Validierung des E-Mail-Verifizierungscodes hinzu.
- Aktualisiere den Benutzerstatus auf 'emailVerified' nach erfolgreicher Verifizierung.
- Setze ein Cookie zur Bestätigung der E-Mail-Verifizierung.

+++ src/features/auth/actions/email-verification.ts
@@ -1,8 +1,10 @@
 "use server";
 
+import { setCookieByKey } from "@/actions/cookies";
 import {
     ActionState,
-    fromErrorToActionState
+    fromErrorToActionState,
+    toActionState
 } from "@/components/form/utils/to-action-state";
 import { createSession } from "@/lib/lucia";
 import { prisma } from "@/lib/prisma";
@@ -10,22 +12,35 @@ import { ticketsPath } from "@/path";
 import { generateRandomToken } from "@/utils/crypto";
 import { redirect } from "next/navigation";
 import { z } from "zod";
+import { getAuthOrRedirect } from "../queries/get-auth-or-redirect";
 import { setSessionCookie } from "../utils/session-cookie";
+import { validateEmailVerificationCode } from "../utils/validate-email-verification-code";
 
 const emailVerficationSchema = z.object({
     code: z.string().length(8),
 });
 
 export const emailVerfication = async (_actionState: ActionState, formData: FormData) => {
+    const { user } = await getAuthOrRedirect();
     try {
         const { code } = emailVerficationSchema.parse(
             Object.fromEntries(formData)
         );
 
-        await prisma.session.deleteMany({
-            where: { userId: '' }
-        })
+        const validCode = await validateEmailVerificationCode(user.id, user.email, code);
 
+        if (!validCode) {
+            return toActionState("ERROR", "Invalid or expired code");
+        }
+
+        await prisma.session.deleteMany({
+            where: { userId: user.id }
+        });
+
+        await prisma.user.update({
+            where: { id: user.id },
+            data: { emailVerified: true }
+        })
 
         const sessionToken = generateRandomToken();
         const session = await createSession(sessionToken, user.id);
@@ -35,5 +50,6 @@ export const emailVerfication = async (_actionState: ActionState, formData: Form
         return fromErrorToActionState(error, formData);
     }
 
+    await setCookieByKey("toast", "Email verified");
     redirect(ticketsPath());
 };
